--- app/useless/login/page.tsx
@@ -1,10 +1,0 @@
-import { LoginForm } from '@/components/signInForm'
-
-export default function Page() {
-  return (
-    <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
-      <div className="w-full max-w-sm">
-      </div>
-    </div>
-  )
-}
