--- app/useless/update-password/page.tsx
@@ -1,11 +1,0 @@
-import { UpdatePasswordForm } from '@/components/update-password-form'
-
-export default function Page() {
-  return (
-    <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
-      <div className="w-full max-w-sm">
-        <UpdatePasswordForm />
-      </div>
-    </div>
-  )
-}
